Move infinity gif into separate file
